main/plugins/org.talend.core.runtime/src/main/java/org/talend/core/runtime/repository/build/TestCaseBuildProvider.java
// ============================================================================
//
// Copyright (C) 2006-2016 Talend Inc. - www.talend.com
//
// This source code is available under agreement available at
// %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
//
// You should have received a copy of the agreement
// along with this program; if not, write to Talend SA
// 9 rue Pages 92150 Suresnes, France
//
// ============================================================================
package org.talend.core.runtime.repository.build;

import java.util.Map;

import org.talend.core.GlobalServiceRegister;
import org.talend.core.model.properties.Item;
import org.talend.core.ui.ITestContainerProviderService;

/**
 * DOC ggu class global comment. Detailled comment
 */
public abstract class TestCaseBuildProvider extends AbstractBuildProvider {

    @Override
    public boolean valid(Map<String, Object> parameters) {
        if (parameters == null || parameters.isEmpty()) {
            return false;
        }
        Object object = parameters.get(ITEM);
        if (object != null && object instanceof Item
                && GlobalServiceRegister.getDefault().isServiceRegistered(ITestContainerProviderService.class)) {
            ITestContainerProviderService testContainerService = (ITestContainerProviderService) GlobalServiceRegister
                    .getDefault().getService(ITestContainerProviderService.class);
            if (testContainerService != null) {
                return testContainerService.isTestContainerItem((Item) object);
            }
        }
        return false;
    }
}
